Added "Show / Hide Notice Board"
Added the Toggle Notice Board on the main page, the news from the main page, like the cards are going to be moving over to a seperate page because they were breaking the responsive view. Capitalized the titles on the hover over links in the Wallet page. Added more images to the links section. [ News moving to seperate page was verified by Lithy ]